refactor: Move code for legacy TS mod deletion to separate file
